requirements.txt
# Web フレームワーク
streamlit>=1.10.0

# 環境変数読み込み
python-dotenv

# OpenAI 埋め込み・モデル呼び出し
openai
tiktoken

# ChromaDB (DuckDB+Parquet ストレージ対応)
chromadb>=0.5.23,<0.6.0

# LangChain 用コミュニティパッケージ
langchain-community
langchain-chroma
langchain-openai

# テキスト分割用
langchain-text-splitters

# データ処理やユーティリティ
numpy
pandas
